[pipelines] Fix logit verification
1. Loading models as a different dtype is no longer supported. As such,
mpnet-bfloat16 was removed.
2. Fix some calls to DistributedRMSNorm that were missed for setting
dtype.

MAX_INTEGRATION_TESTS_REV_ID: cea3d74ea4c1436bc90712049ba2516050e03a3c

+++ max/tests/integration/pipelines/python/deepseekV2/conftest.py
@@ -95,7 +95,7 @@ def dummy_moe_weight(
         config.n_routed_experts if config.n_routed_experts is not None else 64
     )
     return (
-        torch.randn(n_experts, config.hidden_size, dtype=torch.float32)
+        torch.randn(n_experts, config.hidden_size, dtype=torch.bfloat16)
         * WEIGHT_STDDEV
     )
